Move WorkflowManager from src/web-ui/ to src/workflow/
The CLI workflow-command.ts was reaching across into src/web-ui/ to
construct a WorkflowManager for the inspect path. Move WorkflowManager
into src/workflow/ where it belongs layer-wise, and switch its
constructor parameter from `WebEventBus` (concrete) to
`TypedEventBus<WebEventMap>` (the generic + a type-only event-map
import). All other call sites continue to pass `WebEventBus` instances
unchanged because `WebEventBus extends TypedEventBus<WebEventMap>`.

The CLI inspect site now constructs a no-op `TypedEventBus<WebEventMap>`
directly, with no runtime dependency on src/web-ui/.

No behavior change.
